requirements.txt
appdirs==1.4.4
apturl==0.5.2
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
breezy==3.1.0
Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.11
distlib==0.3.1
distro==1.5.0
distro-info==1.0
dnspython==1.16.0
dulwich==0.20.15
ecdsa==0.16.1
fastimport==0.9.8
filelock==3.0.12
GDAL==3.2.2
gpg===1.14.0-unknown
gyp==0.1
html5lib==1.1
httplib2==0.18.1
idna==2.10
importlib-metadata==4.0.1
jeepney==0.7.1
keyring==23.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
louis==3.18.0
lxml==4.6.3
macaroonbakery==1.3.1
more-itertools==4.2.0
netifaces==0.10.9
npm==0.1.1
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
optional-django==0.1.0
paramiko==2.7.2
patiencediff==0.2.1
pexpect==4.8.0
Pillow==8.1.2
pipenv==11.9.0
protobuf==3.12.4
pycairo==1.16.2
pycodestyle==2.6.0
pycryptodomex==3.9.7
pycups==2.0.1
PyGithub==1.43.7
Pygments==2.7.1
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
PyQt5==5.15.4
PyQt5-sip==12.9.0
pyRFC3339==1.1
python-apt==2.2.1
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
python-gitlab==2.5.0
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.25.1
SecretStorage==3.3.1
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
systemd-python==234
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.4.6+ds
virtualenv-clone==0.3.0
wadllib==1.3.5
webencodings==0.5.1
wrapt==1.12.1
xdg==5
xkit==0.0.0
zipp==1.0.0
